Implemented `bitwidth [operand] <width>/off` command
This patch adds a command that enables asar to throw a warning on mismatch of specified bitwidth and operand size.
Example:
to specify that a bitwidth of 16 for a must be used you can do the following.
```
bitwidth a 16
[code]
bitwidth a off
```
xy can be used for x/y width instead of a.
The xy/a field can be omitted and that means that the bitwidth check is enabled for both xy and a.

Changes made:
- Added new warning+error for the implementation of the feature
- Added a variable that holds the current state of the bitwidth checker, currently implemented as a single byte with bit-flags.
- Added a few functions for easy checking of correctness of size
- Added a test case for the new warning.